Make ./config/base always be the settings module
app was referencing a module outside the module root, but other modules
(such as routes/metadata) were still referencing the ./config/base.

Now .config/base acts as an alias to the real settings file, as
specified by the CUDL_SERVICES_CONFIG envar, or ./config/default if
the envar is not provided.