Stop doing userid-specific stuff in the app
There's no need to switch uid in the app itself (we'll never run on a
privileged port). And we shouldn't be managing the ownership of the
cache dirs. As a result, the config no longer contains user or
groupids. bin/www will exit with an error if a user/group is
configured, or the process starts as root to avoid running as an
unexpected user.

This allows the removal of the userid dependency, which uses native
code which no longer builds on modern node versions..