fix: postgres connections failing
The pg library was at 4.x which doesn't seem to work with node 10 - db
connections were failing. 6.x is the most recent version supporting
the old global pg.connect() API which services currently uses. It seems
to work with node 10, so it'll do for now (this codebase is about to
be replaced with a modernised version).