build: Build the release package during the Docker image build
This makes the build self-contained. Previously the Makefile would
build the release package first (on the host machine) and the Docker
build would copy in the built package. This is a bit of a Docker
antipattern.

In addition, the separate build and dev Dockerfiles are now stages in
the main Dockerfile, which allows common elements to be shared between
them, and also allows the node version in use to be specified in one
place.